Option suppression

+++ src/app/service/user.service.ts
@@ -15,12 +15,16 @@ export class UserService {
 
   updateUser(user:User): Observable<User[]>{
     this.httpClient.put("https://127.0.0.1:8000/api/users/"+user.id,user).subscribe();
-    console.log('test');
     return this.getUsers();
   }
 
   deleteUser(user:User): Observable<User[]>{
-    this.httpClient.delete("https://127.0.0.1:8000/api/users/"+user.id).subscribe();
+    if(user.outingsOrganizer == ""){
+      this.httpClient.delete("https://127.0.0.1:8000/api/users/"+user.id).subscribe();
+    }else{
+      user.actif = false;
+      this.updateUser(user);
+    }
     return this.getUsers();
   }
 }
@@ -38,5 +42,6 @@ export interface User{
   actif: boolean,
   campus: string,
   userIdentifier: string
+  outingsOrganizer: string
 }
